+++ src/wirestate-lit/index.ts
@@ -3,7 +3,9 @@
  */
 
 export { injection, InjectionDecorator, InjectionOptions } from "./consumption/injection";
+export { useContainer, UseContainerValue } from "./consumption/use-container";
 export { useInjection, UseInjectionOptions, UseInjectionValue } from "./consumption/use-injection";
+export { useScope, UseScopeValue } from "./consumption/use-scope";
 
 export { onCommand, OnCommandDecorator } from "./commands/on-command";
 export { OnCommandController } from "./commands/on-command-controller";